pyproject.toml
[project]
name = "zed-backup"
version = "0.1.0"
description = "A simple script to backup your Zed config to a github gist"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
